UI corrections in albums and artists In Navigation rather than using the album index or aritists index used the ids which are more reliable and are always unique. Used plurals for song count As album art on device under api 29 maybe an uri so another overload of CommonArtwork is prepared Changes in compose code and some recompositions fixes in AlbumsScreen and ArtistScreen

+++ app/src/main/java/com/sosauce/cutemusic/logic/navigation/Screen.kt
@@ -4,28 +4,24 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 sealed class Screen {
-    @Serializable
-    data object Main : Screen()
+	@Serializable
+	data object Main : Screen()
 
-    @Serializable
-    data object NowPlaying : Screen()
+	@Serializable
+	data object NowPlaying : Screen()
 
-    @Serializable
-    data object Settings : Screen()
+	@Serializable
+	data object Settings : Screen()
 
-    @Serializable
-    data object Albums : Screen()
+	@Serializable
+	data object Albums : Screen()
 
-    @Serializable
-    data object Artists : Screen()
+	@Serializable
+	data object Artists : Screen()
 
-    @Serializable
-    data class AlbumsDetails(
-        val id: Int
-    ) : Screen()
+	@Serializable
+	data class AlbumsDetails(val id: Long) : Screen()
 
-    @Serializable
-    data class ArtistsDetails(
-        val id: Int
-    ) : Screen()
+	@Serializable
+	data class ArtistsDetails(val id: Long) : Screen()
 }
